+++ functions/Actions.php
@@ -434,6 +434,14 @@ $RosarioActions = [];
 	 * @since 11.2
 	 */
 	/*'Food_Service/Students/Accounts.php|table_after' => array(),
+
+	/**
+	 * Food Service tab fields table after
+	 * Add your own fields
+	 *
+	 * @since 11.2
+	 */
+	/*'Food_Service/Student.inc.php|table_after' => array(),
 );*/
 
 /**